chore: unused import remove, code refacot, created decoratos
