+++ src/routes/__init__.py
@@ -5,12 +5,12 @@ from src.routes.memory_info import memory_info_bp
 from src.routes.network_info import network_info_bp
 from src.routes.settings import settings_bp
 from src.routes.speedtest import speedtest_bp
-from src.routes.process import process_bp
+from src.routes.process_info import process_bp
 from src.routes.auth import auth_bp
-from src.routes.network import network_bp
+from src.routes.dashboard_network import network_bp
 from src.routes.other import other_bp
 from src.routes.smtp_email_config import smtp_email_config_bp
-from src.routes.user import user_bp
+from src.routes.user_management import user_management_bp
 from src.routes.graphs import graphs_bp
 from src.routes.ping import ping_bp
 from src.routes.firewall import firewall_bp
@@ -18,3 +18,4 @@ from src.routes.health import health_bp
 from src.routes.api import api_bp
 from src.routes.experimental import experimental_bp
 from src.routes.error_handlers import error_handlers_bp
+from src.routes.profile import profile_bp
