+++ src/templates/settings/control_panel.html
@@ -34,12 +34,12 @@
             </a>
         </li>
         {% endif %}
-        <li class="nav-item"></li>
+        <li class="nav-item">
             <a class="nav-link" id="toggles-tab" data-toggle="tab" href="#feature-toggles" role="tab" aria-controls="feature-toggles" aria-selected="false">
                 <i class="fas fa-toggle-on"></i> Feature
             </a>
         </li>
-        <li class="nav-item"></li>
+        <li class="nav-item">
             <a class="nav-link" id="system-info-tab" data-toggle="tab" href="#system-info" role="tab" aria-controls="system-info" aria-selected="false">
                 <i class="fas fa-info-circle"></i> System Info
             </a>
